+++ apps/docs/src/.vitepress/components/ComponentCard.vue
@@ -30,7 +30,7 @@ const props = defineProps<ComponentCardProps>();
 
 <style lang="scss" scoped>
 .card {
-  border: 1px solid var(--vp-c-default-soft);
+  border: var(--onyx-1px-in-rem) solid var(--vp-c-default-soft);
   border-radius: 0.75rem;
   height: 100%;
   background-color: var(--vp-c-bg);