+++ packages/sit-onyx/src/components/OnyxCheckbox/types.ts
@@ -1,0 +1,22 @@
+export type OnyxCheckboxProps = {
+  /**
+   * Whether the checkbox is checked.
+   */
+  modelValue?: boolean;
+  /**
+   * Label to show.
+   */
+  label?: string;
+  /**
+   * If `true`, an indeterminate indicator is shown.
+   */
+  indeterminate?: boolean;
+  /**
+   * Whether to disable the checkbox and prevent user interaction.
+   */
+  disabled?: boolean;
+  /**
+   * Whether the checkbox is required / has to be checked.
+   */
+  required?: boolean;
+};