+++ packages/sit-onyx/src/components/OnyxIcon/OnyxIcon.stories.ts
@@ -6,9 +6,10 @@ import OnyxIcon from "./OnyxIcon.vue";
 import { ICON_SIZES } from "./types";
 
 const ALL_ICONS = import.meta.glob("../../../node_modules/@sit-onyx/icons/src/assets/*.svg", {
-  as: "raw",
+  query: "?raw",
+  import: "default",
   eager: true,
-});
+}) as Record<string, string>;
 
 /**
  * Mapping between icon SVG content (key) and icon name (value).
@@ -85,7 +86,7 @@ export const Default = {
 export const WithSize = {
   args: {
     ...Default.args,
-    size: "2xl",
+    size: "96px",
   },
 } satisfies Story;