+++ packages/sit-onyx/src/components/OnyxRadioButton/OnyxRadioButton.ct.tsx
@@ -50,13 +50,36 @@ test("should display correctly when disabled", async ({ mount, makeAxeBuilder, p
       name="radio form"
       value="radio-value"
       selected
-      isDisabled
+      disabled
     />,
   );
 
   // ASSERT
   await expect(page.getByRole("radio")).toBeDisabled();
   await expect(page.getByRole("radio")).toBeChecked();
+
+  // ACT
+  const accessibilityScanResults = await makeAxeBuilder().analyze();
+
+  // ASSERT
+  expect(accessibilityScanResults.violations).toEqual([]);
+});
+
+test("should display correctly when invalid", async ({ mount, makeAxeBuilder, page }) => {
+  // ARRANGE
+  await mount(
+    <OnyxRadioButton
+      id="my-id"
+      label="radio label"
+      name="radio form"
+      value="radio-value"
+      errorMessage="radio error"
+      selected
+    />,
+  );
+
+  // ASSERT
+  await expect(page.locator("input:invalid")).toBeAttached();
 
   // ACT
   const accessibilityScanResults = await makeAxeBuilder().analyze();