+++ packages/sit-onyx/src/components/TestInput/TestInput.vue
@@ -167,7 +167,7 @@ watch(
   width: max-content;
   display: inline-block;
   font-family: var(--onyx-font-family);
-  color: var(--onyx-color-text-neutral-intense);
+  color: var(--onyx-color-text-icons-neutral-intense);
 
   &__label {
     margin-right: 8px;