+++ packages/sit-onyx/src/index.ts
@@ -1,5 +1,6 @@
 import "@/styles/index.scss";
 
+export * from "@/components/OnyxCheckbox/types";
 export { default as OnyxHeadline } from "@/components/OnyxHeadline/OnyxHeadline.vue";
 export * from "@/components/OnyxHeadline/types";
 export { default as OnyxIcon } from "@/components/OnyxIcon/OnyxIcon.vue";