+++ apps/alpha-test-app/CHANGELOG.md
@@ -1,5 +1,40 @@
 # alpha-test-app
 
+## 0.0.1-alpha.9
+
+### Patch Changes
+
+- Updated dependencies [95b6e75]
+  - sit-onyx@1.0.0-alpha.12
+
+## 0.0.1-alpha.8
+
+### Patch Changes
+
+- Updated dependencies [97da8df]
+  - sit-onyx@1.0.0-alpha.11
+
+## 0.0.1-alpha.7
+
+### Patch Changes
+
+- Updated dependencies [a4df4af]
+  - sit-onyx@1.0.0-alpha.10
+
+## 0.0.1-alpha.6
+
+### Patch Changes
+
+- Updated dependencies [eda3982]
+  - sit-onyx@0.1.0-alpha.9
+
+## 0.0.1-alpha.5
+
+### Patch Changes
+
+- Updated dependencies [c62476d]
+  - sit-onyx@0.1.0-alpha.8
+
 ## 0.0.1-alpha.4
 
 ### Patch Changes