+++ packages/storybook-utils/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @sit-onyx/storybook-utils
 
+## 1.0.0-alpha.13
+
+### Patch Changes
+
+- Updated dependencies [95b6e75]
+  - sit-onyx@1.0.0-alpha.12
+
+## 1.0.0-alpha.12
+
+### Patch Changes
+
+- Updated dependencies [97da8df]
+  - sit-onyx@1.0.0-alpha.11
+
+## 1.0.0-alpha.11
+
+### Patch Changes
+
+- Updated dependencies [a4df4af]
+  - sit-onyx@1.0.0-alpha.10
+
+## 1.0.0-alpha.10
+
+### Patch Changes
+
+- Updated dependencies [eda3982]
+  - sit-onyx@0.1.0-alpha.9
+
+## 1.0.0-alpha.9
+
+### Patch Changes
+
+- Updated dependencies [c62476d]
+  - sit-onyx@0.1.0-alpha.8
+
 ## 1.0.0-alpha.8
 
 ### Patch Changes